feat: Add study reminders and statistics features
- Implemented StudyReminderRepository for managing study reminders, including create, update, delete, and toggle functionalities.
- Created StudyReminderManager for scheduling notifications and handling reminder alerts.
- Developed RemindersScreen for displaying and managing reminders with a user-friendly interface.
- Added StatisticsScreen to show study statistics, including total study time, lessons completed, and current streak.
- Enhanced HomeScreen to navigate to the new Statistics and Reminders screens.
- Updated QuizRepository and StudyActivityRepository to support new features.
- Added permission handling for notifications on Android 13+.

+++ app/src/main/java/com/sunset/ictstudy/ui/screens/CalendarScreen.kt
@@ -30,7 +30,7 @@ fun CalendarScreen(
     sessions: List<StudySession>,
     onBack: () -> Unit,
     onAddSession: (String, Long, Int) -> Unit,
-    onToggleComplete: (Long, Boolean) -> Unit
+    onToggleComplete: (Int, Boolean) -> Unit
 ) {
     var showAddDialog by remember { mutableStateOf(false) }
     var selectedMonth by remember { mutableStateOf(Calendar.getInstance()) }
